+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Change Log
 
+### v0.0.3 (2015/06/02)
+- Add custom options (showPrefix, fullColorize, parenthesisObject)
+- Fix bug (object cannot be colorized)
+
 ### v0.0.1 (2015/06/02)
 - Init project